backend/requirements.txt
blinker==1.9.0
click==8.2.0
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
flask-login
flask-sqlalchemy
python-dotenv
joblib
openai
pandas
scikit-learn

flask-session
flask-cors
flask-migrate
flask-session
flask-wtf
requests